Refactor: split routers
